Refactor services (#423)
* move transaction filter check in transaction utils

* define getExtraDetails for transactions with queryOptions

* move trimOperations in transaction utils

* remove unused import

* apply transaction logs
