+++ src/endpoints/transactions/transaction.service.ts
@@ -29,6 +29,7 @@ import { CacheInfo } from 'src/common/caching/entities/cache.info';
 import { Constants } from 'src/utils/constants';
 import { GatewayComponentRequest } from 'src/common/gateway/entities/gateway.component.request';
 import { SortOrder } from 'src/common/entities/sort.order';
+import { TransactionUtils } from 'src/utils/transaction.utils';
 
 @Injectable()
 export class TransactionService {
@@ -131,40 +132,12 @@ export class TransactionService {
     return await this.elasticService.getCount('transactions', elasticQuery);
   }
 
-  private isTransactionCountQueryWithAddressOnly(filter: TransactionFilter, address?: string) {
-    if (!address) {
-      return false;
-    }
-
-    const filterToCompareWith: TransactionFilter = {};
-
-    return JSON.stringify(filter) === JSON.stringify(filterToCompareWith);
-  }
-
-  private isTransactionCountQueryWithSenderAndReceiver(filter: TransactionFilter) {
-    if (!filter.sender || !filter.receiver) {
-      return false;
-    }
-
-    if (filter.sender !== filter.receiver) {
-      return false;
-    }
-
-    const filterToCompareWith: TransactionFilter = {
-      sender: filter.sender,
-      receiver: filter.receiver,
-      condition: QueryConditionOptions.should,
-    };
-
-    return JSON.stringify(filter) === JSON.stringify(filterToCompareWith);
-  }
-
   async getTransactionCount(filter: TransactionFilter, address?: string): Promise<number> {
-    if (this.isTransactionCountQueryWithAddressOnly(filter, address)) {
+    if (TransactionUtils.isTransactionCountQueryWithAddressOnly(filter, address)) {
       return this.getTransactionCountForAddress(address ?? '');
     }
 
-    if (this.isTransactionCountQueryWithSenderAndReceiver(filter)) {
+    if (TransactionUtils.isTransactionCountQueryWithSenderAndReceiver(filter)) {
       return this.getTransactionCountForAddress(filter.sender ?? '');
     }
 
@@ -212,62 +185,8 @@ export class TransactionService {
       }
     }
 
-    if (queryOptions && (queryOptions.withScResults || queryOptions.withOperations || queryOptions.withLogs) && elasticTransactions.some(x => x.hasScResults === true)) {
-      // Add scResults to transaction details
-      const elasticQuery = ElasticQuery.create()
-        .withPagination({ from: 0, size: 10000 })
-        .withSort([{ name: 'timestamp', order: ElasticSortOrder.ascending }])
-        .withTerms(new TermsQuery('originalTxHash', elasticTransactions.filter(x => x.hasScResults === true).map(x => x.txHash)));
-
-      const scResults = await this.elasticService.getList('scresults', 'scHash', elasticQuery);
-      for (const scResult of scResults) {
-        scResult.hash = scResult.scHash;
-
-        delete scResult.scHash;
-      }
-
-      const hashes = [...transactions.map((transaction) => transaction.txHash), ...scResults.map((scResult) => scResult.hash)];
-      const logs = await this.transactionGetService.getTransactionLogsFromElastic(hashes);
-
-      const detailedTransactions: TransactionDetailed[] = [];
-      for (const transaction of transactions) {
-        const transactionDetailed = ApiUtils.mergeObjects(new TransactionDetailed(), transaction);
-        const transactionScResults = scResults.filter(({ originalTxHash }) => originalTxHash == transaction.txHash);
-
-        if (queryOptions.withScResults) {
-          transactionDetailed.results = transactionScResults.map(scResult => ApiUtils.mergeObjects(new SmartContractResult(), scResult));
-        }
-
-        if (queryOptions.withLogs) {
-          for (const log of logs) {
-            if (log._id === transactionDetailed.txHash) {
-              transactionDetailed.logs = ApiUtils.mergeObjects(new TransactionLog(), log._source);
-            }
-            else {
-              const foundScResult = transactionDetailed.results.find(({ hash }) => log._id === hash);
-              if (foundScResult) {
-                foundScResult.logs = ApiUtils.mergeObjects(new TransactionLog(), log._source);
-              }
-            }
-          }
-        }
-
-        if (queryOptions.withOperations) {
-          const transactionHashes: string[] = [transactionDetailed.txHash];
-          for (const scResult of transactionScResults) {
-            transactionHashes.push(scResult.hash);
-          }
-          const transactionLogsFromElastic = logs.filter((log) => transactionHashes.includes(log._id));
-          const transactionLogs: TransactionLog[] = transactionLogsFromElastic.map(log => ApiUtils.mergeObjects(new TransactionLog(), log._source));
-          transactionDetailed.operations = await this.tokenTransferService.getOperationsForTransactionLogs(transactionDetailed.txHash, transactionLogs);
-
-          transactionDetailed.operations = this.transactionGetService.trimOperations(transactionDetailed.operations);
-        }
-
-        detailedTransactions.push(transactionDetailed);
-      }
-
-      transactions = detailedTransactions;
+    if (queryOptions && (queryOptions.withScResults || queryOptions.withOperations) && elasticTransactions.some(x => x.hasScResults === true)) {
+      transactions = await this.getExtraDetailsForTransactions(elasticTransactions, transactions, queryOptions);
     }
 
     for (const transaction of transactions) {
@@ -341,5 +260,60 @@ export class TransactionService {
       this.logger.error(`Unhandled error when processing plugin transaction for transaction with hash '${transaction.txHash}'`);
       this.logger.error(error);
     }
+  }
+
+  private async getExtraDetailsForTransactions(elasticTransactions: any[], transactions: Transaction[], queryOptions: TransactionQueryOptions): Promise<TransactionDetailed[]> {
+    const elasticQuery = ElasticQuery.create()
+      .withPagination({ from: 0, size: 10000 })
+      .withSort([{ name: 'timestamp', order: ElasticSortOrder.ascending }])
+      .withTerms(new TermsQuery('originalTxHash', elasticTransactions.filter(x => x.hasScResults === true).map(x => x.txHash)));
+
+    const scResults = await this.elasticService.getList('scresults', 'scHash', elasticQuery);
+    for (const scResult of scResults) {
+      scResult.hash = scResult.scHash;
+
+      delete scResult.scHash;
+    }
+
+    const hashes = [...transactions.map((transaction) => transaction.txHash), ...scResults.map((scResult) => scResult.hash)];
+    const logs = await this.transactionGetService.getTransactionLogsFromElastic(hashes);
+
+    const detailedTransactions: TransactionDetailed[] = [];
+    for (const transaction of transactions) {
+      const transactionDetailed = ApiUtils.mergeObjects(new TransactionDetailed(), transaction);
+      const transactionScResults = scResults.filter(({ originalTxHash }) => originalTxHash == transaction.txHash);
+
+      if (queryOptions.withScResults) {
+        transactionDetailed.results = transactionScResults.map(scResult => ApiUtils.mergeObjects(new SmartContractResult(), scResult));
+      }
+
+      if (queryOptions.withOperations) {
+        const transactionHashes: string[] = [transactionDetailed.txHash];
+        for (const scResult of transactionScResults) {
+          transactionHashes.push(scResult.hash);
+        }
+        const transactionLogsFromElastic = logs.filter((log) => transactionHashes.includes(log._id));
+        const transactionLogs: TransactionLog[] = transactionLogsFromElastic.map(log => ApiUtils.mergeObjects(new TransactionLog(), log._source));
+        transactionDetailed.operations = await this.tokenTransferService.getOperationsForTransactionLogs(transactionDetailed.txHash, transactionLogs);
+
+        transactionDetailed.operations = TransactionUtils.trimOperations(transactionDetailed.operations);
+
+        for (const log of transactionLogsFromElastic) {
+          if (log._id === transactionDetailed.txHash) {
+            transactionDetailed.logs = ApiUtils.mergeObjects(new TransactionLog(), log._source);
+          }
+          else {
+            const foundScResult = transactionDetailed.results.find(({ hash }) => log._id === hash);
+            if (foundScResult) {
+              foundScResult.logs = ApiUtils.mergeObjects(new TransactionLog(), log._source);
+            }
+          }
+        }
+      }
+
+      detailedTransactions.push(transactionDetailed);
+    }
+
+    return detailedTransactions;
   }
 }
